Update for latest Homebrew directory layout (#1)
This module has not be updated in a long time, and Homebrew now has a differentl layout with two git repositories (one for the scripts, like brew, and one for the formulas).

I've tested these changes on my localhost by blowing away /usr/local/Homebrew and they seem to work. The only other change is to change anything that requires Exec["install-homebrew"] to instead require File["/usr/local/bin/brew"], which seems like a better way to externalize it anyways.